working

+++ src/main/java/com/makersacademy/nevermore/model/Cost.java
@@ -22,7 +22,7 @@ public class Cost {
     private Long id;
     private Long user_id;
     private String content;
-    private int price;
+    private Double price;
     private String category;
     public Date date;
 
@@ -30,14 +30,14 @@ public class Cost {
         
     }
 
-    public Cost(String content, Integer price, String category) {
+    public Cost(String content, Double price, String category) {
         this.content = content;
         this.price = price;
         this.category = category;
         
     }
 
-    public Cost(String content, int price, String category, Long user_id, Date date) {
+    public Cost(String content, Double price, String category, Long user_id, Date date) {
         this.content = content;
         this.price = price;
         this.category = category;
@@ -49,7 +49,7 @@ public class Cost {
     //Getters
     public Long getUserid() { return this.user_id; }
     public String getcontent() { return this.content; }
-    public Integer getPrice()  { return this.price;}
+    public Double getPrice()  { return this.price;}
     public String getCategory() { return this.category;}
     public Long getId() { return this.id; }
     public Date getDate() {return this.date;}
@@ -57,7 +57,7 @@ public class Cost {
     //Setters
     public void setUserid(Long user_id) { this.user_id = user_id;}
     public void setContent(String content) { this.content = content; }
-    public void setPrice(Integer price) { this.price = price;}
+    public void setPrice(Double price) { this.price = price;}
     public void setCategory(String category) { this.category = category;}
     public void setDate(Date date) { this.date = date;}
 }
